Rename pyav_extra.py to pyav_hdr.py
